♻️ refactor(config): 重构配置模块拆分工具函数
- 将配置相关工具函数拆分到独立模块
-【重构】从 config-schema.ts 提取工具函数到单独的文件
- 新增 config-builder.ts 处理配置生成逻辑
- 新增 config-keys-generator.ts 处理配置键生成
- 新增 config-metadata-generator.ts 处理元数据生成
- 新增 config-validation.ts 处理配置验证
- 调整主文件导出结构,使用重新导出方式组织模块
