+++ src/config/config-schema.ts
@@ -275,206 +275,15 @@ export type SchemaType = {
 // Generate type
 export type ConfigPath = string; // e.g., "providers.openai.apiKey"
 
-/**
- * Generates configuration keys from schema
- * @param {SchemaType} schema - Configuration schema
- * @param {string} [prefix=''] - Optional prefix for nested keys
- * @returns {Record<string, string>} Generated configuration keys
- */
-export function generateConfigKeys(
-  schema: SchemaType,
-  prefix: string = ""
-): Record<string, string> {
-  const keys: Record<string, string> = {};
-
-  /**
-   * Recursively traverse the configuration object to generate keys
-   * @param {ConfigObject} obj - Current configuration object being processed
-   * @param {string} [path=''] - Current path in the configuration hierarchy
-   */
-  function traverse(obj: ConfigObject, path: string = "") {
-    for (const [key, value] of Object.entries(obj)) {
-      // Generate full path by combining current path and key
-      const fullPath = path ? `${path}.${key}` : key;
-
-      if (isConfigValue(value)) {
-        // For configuration values, generate full configuration key
-        const configKey = fullPath.replace(/\./g, "_").toUpperCase();
-        keys[configKey] = `dish-ai-commit.${fullPath}`;
-      } else {
-        // For nested objects, generate intermediate key and continue traversal
-        const intermediateKey = fullPath.replace(/\./g, "_").toUpperCase();
-        keys[intermediateKey] = `dish-ai-commit.${fullPath}`;
-        traverse(value as ConfigObject, fullPath);
-      }
-    }
-  }
-
-  traverse(schema as unknown as ConfigObject);
-  return keys;
-}
-
-// Add metadata type definition
-/**
- * Configuration metadata item interface
- * @interface ConfigMetadataItem
- */
-export interface ConfigMetadataItem {
-  key: string;
-  defaultValue: any;
-  nested: boolean;
-  parent: string;
-  description: string;
-  type: string;
-  enum?: readonly string[];
-  enumDescriptions?: readonly string[];
-  isSpecial?: boolean;
-}
-
-/**
- * Generates configuration metadata from schema
- * @param {SchemaType} schema - Configuration schema
- * @returns {ConfigMetadataItem[]} Array of metadata items
- */
-export function generateConfigMetadata(
-  schema: SchemaType
-): ConfigMetadataItem[] {
-  const metadata: ConfigMetadataItem[] = [];
-
-  /**
-   * Recursively traverse the configuration object to generate metadata
-   * @param {ConfigObject} obj - Current configuration object being processed
-   * @param {string} [path=''] - Current path in the configuration hierarchy
-   */
-  function traverse(obj: ConfigObject, path: string = "") {
-    for (const [key, value] of Object.entries(obj)) {
-      const fullPath = path ? `${path}.${key}` : key;
-      if (isConfigValue(value)) {
-        const metadataItem: ConfigMetadataItem = {
-          key: fullPath.replace(/\./g, "_").toUpperCase(),
-          defaultValue: value.default,
-          nested: fullPath.includes("."),
-          parent: fullPath.split(".")[0],
-          description: value.description,
-          type: value.type,
-          isSpecial: value.isSpecial,
-        };
-
-        // Only add enum and enumDescriptions if the value is of type ConfigValueTypeString
-        if (value.type === "string" && "enum" in value) {
-          metadataItem.enum = value.enum;
-          metadataItem.enumDescriptions = value.enumDescriptions;
-        }
-
-        metadata.push(metadataItem);
-      } else if (typeof value === "object") {
-        // Handle nested objects
-        traverse(value as ConfigObject, fullPath);
-      }
-    }
-  }
-
-  traverse(schema as unknown as ConfigObject);
-  return metadata;
-}
-
-// Add type checking helper function
-/**
- * Type guard for configuration values
- * @param {unknown} value - Value to check
- * @returns {boolean} Whether value is a ConfigValue
- */
-export function isConfigValue(value: unknown): value is ConfigValue {
-  return (
-    value !== null &&
-    typeof value === "object" &&
-    "type" in value &&
-    "description" in value
-  );
-}
-
-/**
- * Generates configuration object from schema
- * @param {typeof CONFIG_SCHEMA} schema - Configuration schema
- * @param {(key: string) => any} getConfig - Function to retrieve config values
- * @returns {any} Generated configuration object
- */
-export function generateConfiguration(
-  schema: typeof CONFIG_SCHEMA,
-  getConfig: (key: string) => any
-) {
-  const result: any = {};
-
-  /**
-   * Recursively traverses schema to build configuration
-   * @param {ConfigObject} obj - Current configuration object
-   * @param {string} currentPath - Current path in configuration
-   */
-  function traverse(obj: ConfigObject, currentPath: string = "") {
-    for (const [key, value] of Object.entries(obj)) {
-      const newPath = currentPath ? `${currentPath}.${key}` : key;
-
-      if (isConfigValue(value)) {
-        // Get config value or use default
-        const configValue = getConfig(newPath) ?? value.default;
-
-        // Handle nested path
-        const pathParts = newPath.split(".");
-        let current = result;
-
-        // Build nested object structure
-        for (let i = 0; i < pathParts.length - 1; i++) {
-          if (!(pathParts[i] in current)) {
-            current[pathParts[i]] = {};
-          }
-          current = current[pathParts[i]];
-        }
-
-        // Set final value
-        current[pathParts[pathParts.length - 1]] = configValue;
-      } else {
-        // Continue traversing nested objects
-        traverse(value as ConfigObject, newPath);
-      }
-    }
-  }
-
-  traverse(schema as unknown as ConfigObject);
-  return result;
-}
-
-/**
- * Gets all configuration paths from schema
- * @param {typeof CONFIG_SCHEMA} schema - Configuration schema
- * @returns {string[]} Array of all configuration paths
- */
-export function getAllConfigPaths(schema: typeof CONFIG_SCHEMA): string[] {
-  const paths: string[] = [];
-
-  function traverse(obj: ConfigObject, currentPath: string = ""): void {
-    for (const [key, value] of Object.entries(obj)) {
-      const newPath = currentPath ? `${currentPath}.${key}` : key;
-      if (isConfigValue(value)) {
-        paths.push(newPath);
-      } else {
-        traverse(value as ConfigObject, newPath);
-      }
-    }
-  }
-
-  traverse(schema as unknown as ConfigObject);
-  return paths;
-}
-
-/**
- * Gets configuration paths for a specific category
- * @param {typeof CONFIG_SCHEMA} schema - Configuration schema
- * @param {keyof typeof CONFIG_SCHEMA} category - Category to get paths for
- * @returns {string[]} Array of category configuration paths
- */
-export function getCategoryConfigPaths(
-  schema: typeof CONFIG_SCHEMA,
-  category: keyof typeof CONFIG_SCHEMA
-): string[] {
-  return getAllConfigPaths(schema).filter((path) => path.startsWith(category));
-}
+// Re-export utility functions from new modules
+export { generateConfigKeys } from "./utils/config-keys-generator";
+export {
+  generateConfigMetadata,
+  type ConfigMetadataItem,
+} from "./utils/config-metadata-generator";
+export { isConfigValue } from "./utils/config-validation";
+export {
+  generateConfiguration,
+  getAllConfigPaths,
+  getCategoryConfigPaths,
+} from "./utils/config-builder";
